Refactor suffix tree implementation
This commit is an almost complete rewrite of the suffix tree
implementation. It tries to solve multiple issues present in the
original implementation.

Firstly, this refactor tries to improve data locality (of reference) by
having multiple storage pools than one diskmap. This relays the problem
of data locality to the underlying file system, and we trust that
developers of those file systems are more familiar with this topic. This
also makes data more aligned. For example, there is no longer need to
store extra metadata together with the string, so the file system can
group it in perfect 4K blocks, rathar than (4K + sizeof(metadata))
blocks. As an extra benefit, the linux kernel seems to do more caching.

Another benefit of having multiple storage pools is that although
defragmentation is not implemented yet, it can be more easily
implemented in the future, as all records are now of equal size.

Secondly, this refactor tries to address the readability issue, by
having multiple typedefs and introducing composable functions. The macro
_composable is defined as static inline __attribute__((always_inline)).
This means that composable functions are of zero overhead while
significantly improving readability. This also makes debugging much
easier, as the _composable macro is turned off for debug builds.

The introduction of composable macros also serves to address the
criticism that the original implementations using too many macros. To
make things worse, those macros are not self contained, that is, they
makes additional assumptions such as the functions containing them
having an argument diskmap or having an argument tree. Some macros are
still used in the new implementation due to the impossibility of
replacing them with equivalent functions, but they are now all self
contained.

Note that suffix_tree.c after this commit is not yet ready to use, or in
any way usable. The refactor is still in progress, and there is
currently minimal comments. This issue should no longer present after
the entire refactoring process is finished.

This commit breaks compatibility. A migration is possible, but will not
be published by us, as currently this extension has few, if not none, users.