Traverse tree instead of pile up tags
This is a very rough draft of changing the tagging strategy. Instead of
piling up children tags at parent node, we instead just traverse the
tree to obtain those tags.

Originally, we did not choose to do so due to concerns in performance of
tree traversal. However, the disk turned out to be a much more severe
problem and the additional overhead is actually minor.